+++ platform/src/codatplatform/models/operations/get_company_data_status.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import datastatus as shared_datastatus
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import datastatus as shared_datastatus
 from typing import Dict, Optional
 
 
@@ -20,13 +19,11 @@ class GetCompanyDataStatusRequest:
 class GetCompanyDataStatusResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     data_statuses: Optional[Dict[str, shared_datastatus.DataStatus]] = dataclasses.field(default=None)
     r"""OK"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""